+++ addons/fancyholograms/src/main/java/me/m56738/easyarmorstands/fancyholograms/element/HologramElement.java
@@ -1,6 +1,7 @@
 package me.m56738.easyarmorstands.fancyholograms.element;
 
 import de.oliver.fancyholograms.api.HologramManager;
+import de.oliver.fancyholograms.api.data.BlockHologramData;
 import de.oliver.fancyholograms.api.hologram.Hologram;
 import me.m56738.easyarmorstands.EasyArmorStandsPlugin;
 import me.m56738.easyarmorstands.api.editor.Session;
@@ -17,6 +18,7 @@ import me.m56738.easyarmorstands.api.property.PropertyRegistry;
 import me.m56738.easyarmorstands.api.util.BoundingBox;
 import me.m56738.easyarmorstands.editor.EntityPositionProvider;
 import me.m56738.easyarmorstands.editor.EntityRotationProvider;
+import me.m56738.easyarmorstands.editor.OffsetProvider;
 import me.m56738.easyarmorstands.fancyholograms.FancyHologramsAddon;
 import me.m56738.easyarmorstands.fancyholograms.editor.layer.HologramRootLayer;
 import me.m56738.easyarmorstands.permission.Permissions;
@@ -57,7 +59,7 @@ public class HologramElement implements SelectableElement, DestroyableElement, E
     @Override
     public @NotNull Button createButton(@NotNull Session session) {
         return new PointButton(session,
-                new EntityPositionProvider(properties),
+                new EntityPositionProvider(properties, getOffsetProvider(properties)),
                 new EntityRotationProvider(properties));
     }
 
@@ -68,7 +70,15 @@ public class HologramElement implements SelectableElement, DestroyableElement, E
 
     @Override
     public @NotNull ToolProvider getTools(@NotNull PropertyContainer properties) {
-        return new HologramToolProvider(properties);
+        return new HologramToolProvider(properties, getOffsetProvider(properties));
+    }
+
+    private OffsetProvider getOffsetProvider(PropertyContainer properties) {
+        if (hologram.getData() instanceof BlockHologramData) {
+            return new BlockHologramOffsetProvider(properties);
+        } else {
+            return OffsetProvider.zero();
+        }
     }
 
     @Override