+++ addons/fancyholograms/src/main/java/me/m56738/easyarmorstands/fancyholograms/element/HologramToolProvider.java
@@ -5,6 +5,8 @@ import me.m56738.easyarmorstands.api.editor.tool.AxisScaleTool;
 import me.m56738.easyarmorstands.api.editor.tool.ScaleTool;
 import me.m56738.easyarmorstands.api.editor.tool.ToolContext;
 import me.m56738.easyarmorstands.api.property.PropertyContainer;
+import me.m56738.easyarmorstands.editor.EntityPositionProvider;
+import me.m56738.easyarmorstands.editor.OffsetProvider;
 import me.m56738.easyarmorstands.editor.display.tool.DisplayAxisScaleTool;
 import me.m56738.easyarmorstands.editor.display.tool.DisplayScaleTool;
 import me.m56738.easyarmorstands.element.SimpleEntityToolProvider;
@@ -12,8 +14,9 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public class HologramToolProvider extends SimpleEntityToolProvider {
-    public HologramToolProvider(PropertyContainer properties) {
+    public HologramToolProvider(PropertyContainer properties, OffsetProvider offsetProvider) {
         super(properties);
+        positionProvider = new EntityPositionProvider(properties, offsetProvider);
     }
 
     @Override
@@ -23,7 +26,7 @@ public class HologramToolProvider extends SimpleEntityToolProvider {
 
     @Override
     public @Nullable AxisScaleTool scale(@NotNull ToolContext context, @NotNull Axis axis) {
-        if (context.position() == position() && context.rotation() == rotation()) {
+        if (context.rotation() == rotation()) {
             return new DisplayAxisScaleTool(context, properties, axis);
         }
         return super.scale(context, axis);