+++ src/main/java/me/m56738/easyarmorstands/element/DisplayToolProvider.java
@@ -7,24 +7,24 @@ import me.m56738.easyarmorstands.api.editor.tool.ScaleTool;
 import me.m56738.easyarmorstands.api.editor.tool.ToolContext;
 import me.m56738.easyarmorstands.api.editor.tool.ToolProvider;
 import me.m56738.easyarmorstands.api.property.PropertyContainer;
-import me.m56738.easyarmorstands.api.util.RotationProvider;
 import me.m56738.easyarmorstands.api.property.type.DisplayPropertyTypes;
-import me.m56738.easyarmorstands.editor.display.DisplayOffsetProvider;
+import me.m56738.easyarmorstands.api.util.RotationProvider;
+import me.m56738.easyarmorstands.editor.EntityPositionProvider;
+import me.m56738.easyarmorstands.editor.EntityRotationProvider;
+import me.m56738.easyarmorstands.editor.OffsetProvider;
 import me.m56738.easyarmorstands.editor.display.DisplayRotationProvider;
 import me.m56738.easyarmorstands.editor.display.tool.DisplayAxisRotateTool;
 import me.m56738.easyarmorstands.editor.display.tool.DisplayAxisScaleTool;
 import me.m56738.easyarmorstands.editor.display.tool.DisplayScaleTool;
-import me.m56738.easyarmorstands.editor.EntityPositionProvider;
-import me.m56738.easyarmorstands.editor.EntityRotationProvider;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public class DisplayToolProvider extends SimpleEntityToolProvider implements ToolProvider {
     private final RotationProvider entityRotationProvider;
 
-    public DisplayToolProvider(PropertyContainer properties) {
+    public DisplayToolProvider(PropertyContainer properties, OffsetProvider offsetProvider) {
         super(properties);
-        positionProvider = new EntityPositionProvider(properties, new DisplayOffsetProvider(properties));
+        positionProvider = new EntityPositionProvider(properties, offsetProvider);
         rotationProvider = new DisplayRotationProvider(properties);
         entityRotationProvider = new EntityRotationProvider(properties);
     }
@@ -45,7 +45,7 @@ public class DisplayToolProvider extends SimpleEntityToolProvider implements Too
 
     @Override
     public @Nullable AxisScaleTool scale(@NotNull ToolContext context, @NotNull Axis axis) {
-        if (context.position() == position() && context.rotation() == rotation()) {
+        if (context.rotation() == rotation()) {
             return new DisplayAxisScaleTool(context, properties, axis);
         }
         return super.scale(context, axis);