+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Improvements
+
+* Move block display pivot to the center
+
 # Fixes
 
 * Fix being able to edit unloaded entities